+++ src/settingsObjects/deviceSettings/sonosDeviceSettings.ts
@@ -7,9 +7,23 @@ export class SonosDeviceSettings extends DeviceSettings {
    * @default 80
    */
   public maxPlayOnAllVolume: number = 80;
+  /**
+   * The default volume to inform of non critical stuff during daytime
+   * @type {number}
+   * @default 80
+   */
+  public defaultDayAnounceVolume: number = 80;
+  /**
+   * The default volume to inform of non critical stuff during nighttime
+   * @type {number}
+   * @default 80
+   */
+  public defaultNightAnounceVolume: number = 40;
 
   public fromPartialObject(data: Partial<SonosDeviceSettings>): void {
     this.maxPlayOnAllVolume = data.maxPlayOnAllVolume ?? this.maxPlayOnAllVolume;
+    this.defaultDayAnounceVolume = data.defaultDayAnounceVolume ?? this.defaultDayAnounceVolume;
+    this.defaultNightAnounceVolume = data.defaultNightAnounceVolume ?? this.defaultNightAnounceVolume;
     super.fromPartialObject(data);
   }